time mgmt: complete the debt-bank fix (clock/sprt/effort all signed)
The uncommitted core-bank fix (bank int64, debt clawback) from the SPRT
worktree, now on main: BankedClock (bridge/gauntlet shipped path), sprt
effortBank (the +34 measurement path), and mirror EffortBank all conserve
total budget. Honest re-SPRT +34 +/- 32 (supersedes the +51 confound).

+++ internal/ucibridge/adaptive_test.go
@@ -69,7 +69,7 @@ func TestBridgeAdaptive(t *testing.T) {
 	if b.clock == nil {
 		t.Fatal("adaptive mode did not create the BankedClock")
 	}
-	if bank := b.clock.Bank(); bank > 8*b.clock.Base {
+	if bank := b.clock.Bank(); bank > int64(8*b.clock.Base) {
 		t.Errorf("bank %d exceeds cap %d", bank, 8*b.clock.Base)
 	}
 	t.Logf("adaptive bridge: 3 legal moves; final bank=%d (base=%d)", b.clock.Bank(), b.clock.Base)
